refactor(gui): split digital modes + SWR sweep into their own TUs (#3351 Phase 1e)
Sixth PR of the #3351 monolith-decomposition series. Pure code motion —
no behavior change, no MainWindow.h changes. Two new TUs in one PR
because the source region was one contiguous block spanning both
concerns.

MainWindow_DigitalModes.cpp (1,195 lines):
- RADE digital voice: activateRADE (311 lines) / deactivateRADE /
  onRadeSliceModeChanged
- Classic FreeDV: activateFdvDisplay / deactivateFdvDisplay / meter
  routing / FreeDV Reporter window / spot reporting start/stop
- DAX: startDax (201) / stopDax / wireDaxSlice / onDaxChannelChanged
- AX.25 / AetherModem: decode dialog + KISS TNC startup
- RTTY: decoder output routing

MainWindow_SwrSweep.cpp (617 lines):
- The complete AetherSweep lifecycle: input locking, plot management,
  RF stepping, the sweep state machine (start / advance / finish /
  tune-stop handoff), pan overlay.

Supporting moves, called out for review:

- kSwrSweep* constants (13) promoted to MainWindowHelpers.h — the
  constructor's poll-timer setup in MainWindow.cpp shares them with the
  sweep state machine.
- One #ifdef HAVE_RADE guard re-seated: the cut boundary between the
  two new TUs fell between an #ifdef (tail of the SWR region) and the
  code it guarded (head of the RADE region). The guard now opens in
  MainWindow_DigitalModes.cpp directly before activateRADE. Conditional
  nesting verified balanced in both TUs (awk depth check = 0).
- One section comment that straddled the original cut boundary restored
  to MainWindow.cpp with the applet-panel code it describes.

Breadcrumbs at all four excision sites.

MainWindow.cpp: 13,095 → 11,362 lines (series: 19,474 → 11,362, −41.7%).
Verified: full build clean, all 33 ctest suites pass, a11y lint clean.

Refs #3351. Principle XI.
