+++ CMakeLists.txt
@@ -638,6 +638,8 @@ set(GUI_SOURCES
     src/gui/MainWindow_Controllers.cpp
     src/gui/MainWindow_Menus.cpp
     src/gui/MainWindow_Shortcuts.cpp
+    src/gui/MainWindow_DigitalModes.cpp
+    src/gui/MainWindow_SwrSweep.cpp
     src/gui/MainWindow_Wiring.cpp
     src/gui/AgcCalibrationDialog.cpp
     src/gui/AudioDeviceChangeDialog.cpp
